Split virtual methods diff test in two files
