+++ test/FirstDerivative/VirtualMethodsCall.C
@@ -6,17 +6,7 @@
 
 extern "C" int printf(const char* fmt, ...);
 
-// Overload
-
-float func(float x) {
-  return x;
-}
-
-double func(double x) {
-  return x+x;
-}
-
-// Classes: forward, hide, overload, virtual and overriden methods
+// Classes: forward, hide, virtual and overriden methods
 
 class A {
 public:
@@ -27,16 +17,6 @@ public:
 
   float f(float x) {
     return x;
-  }
-
-  // Overload
-
-  float f1(float x) {
-    return x+x+x;
-  }
-
-  double f1(double x) {
-    return x+x+x+x;
   }
 
   // Virtual/Override
@@ -87,7 +67,6 @@ float A::m(float x, float y) {
   //return vm_darg0_cf.execute(*this, x, y) + vm_darg1_cf.execute(*this, x, y);
 }
 
-
 class B : public A {
 public:
   B() {}
@@ -97,12 +76,6 @@ public:
 
   float f(float x) {
     return x*x;
-  }
-
-  // Overload
-
-  float f1(float x) {
-    return x+x+x+x+x;
   }
 
   // Virtual/Override
@@ -147,12 +120,6 @@ public:
     return x*x*x;
   }
 
-  // Overload
-
-  double f1(double x) {
-    return x+x+x+x+x+x;
-  }
-
   // Virtual/Override
 
   float vm(float x, float y) override {
@@ -172,22 +139,6 @@ int main () {
   A a;
   B b;
   B1 b1;
-
-  // functions overload
-  auto func_darg0_float = clad::differentiate(static_cast<float(*)(float)>(&func), 0);
-//CHECK: float func_darg0(float x) {
-//CHECK-NEXT:    float _d_x = 1;
-//CHECK-NEXT:    return _d_x;
-//CHECK-NEXT: }
-  auto func_darg0_double = clad::differentiate((double(*)(double))&func, 0);
-//CHECK: double func_darg0(double x) {
-//CHECK-NEXT:    double _d_x = 1;
-//CHECK-NEXT:    return _d_x + _d_x;
-//CHECK-NEXT: }
-  printf("Result is = %f\n", func_darg0_float.execute(2.0)); // CHECK-EXEC: Result is = 1.0000
-  printf("Result is = %f\n", func_darg0_double.execute(2.0)); // CHECK-EXEC: Result is = 2.0000
-
-  printf("---\n"); // CHECK-EXEC: ---
 
   //
   printf("Result is = %f\n", a.f(2.0)); // CHECK-EXEC: Result is = 2.0000
@@ -215,39 +166,6 @@ int main () {
   printf("Result is = %f\n", f_darg0_A.execute(a, 2.0)); // CHECK-EXEC: Result is = 1.0000
   printf("Result is = %f\n", f_darg0_B.execute(b, 2.0)); // CHECK-EXEC: Result is = 4.0000
   printf("Result is = %f\n", f_darg0_B1.execute(b1, 2.0)); // CHECK-EXEC: Result is = 12.0000
-
-  printf("---\n"); // CHECK-EXEC: ---
-
-  auto f1_darg0_float_A = clad::differentiate((float(A::*)(float))&A::f1, 0);
-//CHECK: float f1_darg0(float x) {
-//CHECK-NEXT:     float _d_x = 1;
-//CHECK-NEXT:     return _d_x + _d_x + _d_x;
-//CHECK-NEXT: }
-  auto f1_darg0_double_A = clad::differentiate((double(A::*)(double))&A::f1, 0);
-//CHECK: double f1_darg0(double x) {
-//CHECK-NEXT:     double _d_x = 1;
-//CHECK-NEXT:     return _d_x + _d_x + _d_x + _d_x;
-//CHECK-NEXT: }
-  auto f1_darg0_float_B = clad::differentiate((float(B::*)(float))&B::f1, 0);
-//CHECK: float f1_darg0(float x) {
-//CHECK-NEXT:     float _d_x = 1;
-//CHECK-NEXT:     return _d_x + _d_x + _d_x + _d_x + _d_x;
-//CHECK-NEXT: }
-  auto f1_darg0_double_B = clad::differentiate((double(B::*)(double))&B::f1, 0); // resolve to float(B::*)(float)
-//
-  auto f1_darg0_float_B1 = clad::differentiate((float(B1::*)(float))&B1::f1, 0); // resolve to double(B1::*)(double)
-//
-  auto f1_darg0_double_B1 = clad::differentiate((double(B1::*)(double))&B1::f1, 0);
-//CHECK: double f1_darg0(double x) {
-//CHECK-NEXT:     double _d_x = 1;
-//CHECK-NEXT:     return _d_x + _d_x + _d_x + _d_x + _d_x + _d_x;
-//CHECK-NEXT: }
-  printf("Result is = %f\n", f1_darg0_float_A.execute(a, 2.0)); // CHECK-EXEC: Result is = 3.0000
-  printf("Result is = %f\n", f1_darg0_double_A.execute(a, 2.0)); // CHECK-EXEC: Result is = 4.0000
-  printf("Result is = %f\n", f1_darg0_float_B.execute(b, 2.0)); // CHECK-EXEC: Result is = 5.0000
-  printf("Result is %s\n", f1_darg0_double_B.execute(b, 2.0)<1 ? "float" : "other"); // CHECK-EXEC: Result is float
-  printf("Result is %s\n", f1_darg0_float_B1.execute(b1, 2.0f)<1 ? "double" : "other"); // CHECK-EXEC: Result is double
-  printf("Result is = %f\n", f1_darg0_double_B1.execute(b1, 2.0)); // CHECK-EXEC: Result is = 6.0000
 
   printf("---\n"); // CHECK-EXEC: ---
